Fix compilation in VC++ due to a stupid #define NEAR, which seems to be required - worked around that by renaming math::near to math::isNear. Silence C++17 deprecation warnings in Eigen library, and a few more number conversion warnings.

+++ libs/math/AABB.h
@@ -223,7 +223,7 @@ inline const Vector3& AABB::getExtents() const
 
 inline float AABB::getRadius() const
 {
-	return extents.getLength(); // Pythagorean length of extents vector
+	return static_cast<float>(extents.getLength()); // Pythagorean length of extents vector
 }
 
 inline bool AABB::contains(const AABB& other) const
